+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.4.0] - 2020-06-17
+### Removed
+- `consul.base` **Checks.compat** as it was marked as deprecated.
+
+### Added
+- `consul.base` *token* argument for **Agent.leave**, **Agent.force_leave** and **Event.list**.
+
+### Fixed
+- `consul.base` some lint work.
+
 ## [1.3.0] - 2019-06-27
 ### Changed
 - Package name **consul** => **consul-reborn**. Because name **Consul** is occupied at PyPi.